prompt for driver and plugin install

+++ src/serverInstall.ts
@@ -7,6 +7,7 @@ import { plugins } from './plugin.js';
 const ui = new inquirer.ui.BottomBar();
 // @ts-ignore
 import shelljs from 'shelljs';
+import chalk from 'chalk';
 
 export async function installAppiumServer() {
   newLine();
@@ -14,7 +15,7 @@ export async function installAppiumServer() {
   try {
     await run('npm install -g appium@next');
     const { stdout } = await run('appium -v');
-    spinner.succeed(`Successfully installed server version ${stdout}`);
+    spinner.succeed(`💥 💥 💥 Successfully installed server version ${stdout}`);
   } catch (err: any) {
     spinner.fail(err);
     throw new Error(err);
@@ -38,8 +39,8 @@ export async function getDriver() {
     spinner.succeed();
     return drivers;
   } catch (err: any) {
-    spinner.fail();
-    throw new Error(err);
+    spinner.fail(err);
+    spinner.stop();
   } finally {
     spinner.stop();
   }
@@ -70,9 +71,28 @@ export async function installPlugin() {
       choices: ['npm', 'github', 'git', 'local'],
     }
   ]);
+  const { pluginPath } = await inquirer
+      .prompt([
+        {
+          name: 'pluginPath',
+          message: 'Source of plugin',
+        },
+      ])
+  const installedPlugins = await shelljs.exec('appium plugin list --installed --json', { silent: true });
+  const pluginNamesInstalled: any = Object.values(JSON.parse(installedPlugins.stdout)).map((p: any) => p.pkgName);
   await Promise.all(
       requiredPlugins.plugins.map(async (pluginName: string) => {
-        await shelljs.exec(`appium plugin install --source ${source} ${pluginName}`);
+        ui.log.write('Checking if plugin is already installed');
+        if(pluginNamesInstalled.includes(pluginName)) {
+            ui.log.write(chalk.yellow(`💾 💾 Plugin ${pluginName} already installed`));
+            return;
+          } else {
+          if (pluginPath === 'npm') {
+            await shelljs.exec(`appium plugin install --source ${source} ${pluginName}`);
+          } else {
+            await shelljs.exec(`appium plugin install --source ${source} --package ${pluginPath} plugin`);
+          }
+        }
       })
   );
